+++ packages/serverless-components/aws-cloudfront/__tests__/cache-behavior-options.test.js
@@ -23,7 +23,9 @@ describe("Input origin as a custom url", () => {
   it("creates distribution with custom default behavior options", async () => {
     await component.default({
       defaults: {
-        ttl: 0,
+        minTTL: 0,
+        defaultTTL: 0,
+        maxTTL: 31536000,
         forward: {
           headers: ["Accept", "Accept-Language"],
           cookies: "all",
@@ -53,14 +55,18 @@ describe("Input origin as a custom url", () => {
   it("creates distribution with custom behavior options", async () => {
     await component.default({
       defaults: {
-        ttl: 0
+        minTTL: 0,
+        defaultTTL: 0,
+        maxTTL: 31536000
       },
       origins: [
         {
           url: "https://mycustomorigin.com",
           pathPatterns: {
             "/sample/path": {
-              ttl: 0,
+              minTTL: 0,
+              defaultTTL: 0,
+              maxTTL: 0,
               forward: {
                 headers: "all",
                 cookies: ["auth-token"],